requirements.txt
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
loguru==0.6.0
